+++ test/services/forest_liana/value_stat_getter_test.rb
@@ -17,7 +17,7 @@ module ForestLiana
       assert stat.record.value[:countCurrent] == 0
     end
 
-    test 'Value stat getter with a filter on a belongs_to field' do
+    test 'Value stat getter with a filter on a belongs_to integer field' do
       stat = ValueStatGetter.new(BelongsToField, {
         type: "Value",
         collection: "belongs_to_field",
@@ -27,6 +27,40 @@ module ForestLiana
         filters: [{
           field: "has_one_field:id",
           value: "3"
+        }]
+      })
+
+      stat.perform
+      assert stat.record.value[:countCurrent] == 1
+    end
+
+    test 'Value stat getter with a filter on a belongs_to boolean field' do
+      stat = ValueStatGetter.new(BelongsToField, {
+        type: "Value",
+        collection: "belongs_to_field",
+        timezone: "+02:00",
+        aggregate: "Count",
+        filterType: "and",
+        filters: [{
+          field: "has_one_field:checked",
+          value: "false"
+        }]
+      })
+
+      stat.perform
+      assert stat.record.value[:countCurrent] == 0
+    end
+
+    test 'Value stat getter with a filter on a belongs_to enum field' do
+      stat = ValueStatGetter.new(BelongsToField, {
+        type: "Value",
+        collection: "belongs_to_field",
+        timezone: "+02:00",
+        aggregate: "Count",
+        filterType: "and",
+        filters: [{
+          field: "has_one_field:status",
+          value: "pending"
         }]
       })